rename: 파일 디렉토리 변경에 따른 종속성 경로 변경(#5)
