+++ SSoc-Client/src/components/Icons/TabIcon.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { View } from "react-native";
 import { Badge } from "./Badge";
 import { Icon } from "./Icons";
