+++ SSoc-Client/src/components/header/HeaderTitle.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Typography } from "../Typography";
+import { Typography } from "../Basic/Typography";
 
 export const HeaderTitle = (props) => {
   return <Typography fontSize={18}>{props.title}</Typography>;
